Fix poetry install issue on platform.sh.
`poetry install` raised: `AttributeError: 'Link' object has no attribute
'is_absolute'` during build on platform.sh.

It appears that it's possible for poetry-core to update even if poetry
is locked. Our issue wasa caused by a regression introduced in 1.0.4.

Updated poetry following the advice in the linked github issue and
updated pip to fix an issue selecting the correct wheel for cryptography
`36.0.0` on platform.sh.

Locked Poetry to `1.1.12` and pip to `21.3.1` for docker and platform.sh

Related issues:

 - [Link object being passed instead of Path to poetry.core.packages.file_dependency.FileDependency in poetry-core 1.0.5](h
 - [fix: Broken cache on Windows #4531](https://github.com/python-poetry/poetry/pull/4531)